+++ src/Core/Resolver/EntityInventoryItemsResolver.php
@@ -13,6 +13,7 @@ namespace Velkuns\GameTextEngine\Core\Resolver;
 
 use Velkuns\GameTextEngine\Rpg\Entity\EntityInterface;
 use Velkuns\GameTextEngine\Rpg\Entity\EntityInventory;
+use Velkuns\GameTextEngine\Rpg\Modifier\Modifier;
 
 readonly class EntityInventoryItemsResolver implements ResolverInterface
 {
@@ -24,5 +25,14 @@ readonly class EntityInventoryItemsResolver implements ResolverInterface
     public function resolve(string $type, EntityInterface $entity): EntityInventory
     {
         return $entity->getInventory();
+    }
+
+    /**
+     * @param list<Modifier> $modifiers
+     * @codeCoverageIgnore
+     */
+    public function resolveValue(string $type, EntityInterface $entity, array $modifiers = []): int
+    {
+        return 0;
     }
 }